requirements.txt
asgiref==3.4.1
Django==3.2.7
pytz==2021.1
sqlparse==0.4.1
djangorestframework
psycopg2==2.8.5
python-dotenv
django-cors-headers
djangorestframework-simplejwt
django-taggit
django-s3direct
whitenoise==5.3.0
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==8.3.2